Day-13/restaurant-AI-assistant/pyproject.toml
[project]
name = "blob-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.0",
    "langchain>=1.2.3",
    "langchain-google-genai>=4.1.3",
    "langgraph-checkpoint-sqlite>=3.0.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
